+++ docs/changelog.md
@@ -12,6 +12,7 @@ Sprint 7 / February 29, 2020
 - added Timeward site heading to header.
 - added Date component to be used as singular date objects inside Timeline component.
 - added tests for Button component.
+- update Header nav link styles for hover effects.
 
 ## Release 0.6.0
 Sprint 6 / February 15, 2020